Fix copy relative path command when file is not relative to workspace
The asRelativePath(...) method returns the path without a leading '/'
even when returning an absolute path because the file is not in the
workspace.

resource.path (used for Copy Path commands) does include a leading '/'.

This leads to inconsistency with the Copy Relative Path command when the
file is not within the workspace.  To fix this, modify the regex to make
the leading '/' optional